requirements.txt
web3==5.4.0
py-solc==3.2.0